… tests/test_performance.py
… voicereel/PERFORMANCE.md
… voicereel/performance_optimizer.py
… voicereel/tasks.py
… voicereel/todo.md